Processor SDK 12_00_00_01 CICD Release

+++ mcu_plus_sdk/tools/boot/uart_bootloader.py
@@ -3,6 +3,8 @@ import sys
 import getopt
 import time
 import shlex
+import string
+import struct
 
 try:
     import serial
@@ -24,6 +26,10 @@ USAGE: python uart_bootloader.py [OPTIONS]
 -b, --bootloader=    Path to the UART Bootloader binary
 
 -f, --file=          Path to the appimage binary
+
+--cfg=               Path to the configuration file with the entries of the two flags given above, namely bootloader and file.
+                     This file can be used as an alternative to passing the two flags given above.
+                     Use either the --cfg flag or the --bootloader and --file flags.
 '''
 
 BOOTLOADER_UART_STATUS_LOAD_SUCCESS            = 0x53554343
@@ -32,6 +38,10 @@ BOOTLOADER_UART_STATUS_APPIMAGE_SIZE_EXCEEDED  = 0x45584344
 
 BOOTLOADER_END_OF_FILES_TRANSFER               = 0x454F4654
 BOOTLOADER_END_OF_FILES_TRANSFER_WORD_LENGTH   = 4
+
+BOOTLOADER_UART_MAGIC                          = 0xBF0000BF
+BOOTLOADER_UART_FILE_RECEIVE_COMPLETE          = 0xBF000002
+BOOTLOADER_UART_SEND_FILE                      = 0xBF000003
 
 mySerPort = "No Serial Port Chosen"
 myBaudRate = 115200
@@ -65,6 +75,171 @@ def parse_response_evm(filename):
         status = "[STATUS] ERROR: Bad response from EVM !!!"
 
     return status
+
+# Parse message requests for program segments from EVM
+def parse_msg_response(resp_bytes):
+    '''
+    /* UART protocol request header */
+    typedef struct Bootloader_UartProtocolReq_s
+    {
+        uint64_t magic;
+        uint64_t virtMemOffset;
+        uint64_t cmd;
+        uint64_t len;
+
+    } Bootloader_UartProtocolReq;
+    '''
+
+    uart_header_str = "<QQQQ"
+
+    status = -1
+
+    magic, virt_mem_offset, cmd , req_length = struct.unpack(uart_header_str, resp_bytes)
+
+    if(magic == BOOTLOADER_UART_MAGIC):
+        status = 0
+    else:
+        status = -1
+
+    return status, cmd, virt_mem_offset, req_length
+
+# Send the file to EVM
+def xmodem_send_file(port, baudrate, timeout, f_stream, show_progress):
+
+    status = None
+    open_serial_port();
+
+    f_stream.seek(0, os.SEEK_END)
+    bar_filesize = f_stream.tell()
+    f_stream.seek(0, os.SEEK_SET)
+
+    if show_progress:
+        bar = tqdm(total=bar_filesize, unit="bytes", leave=True, desc="Sending file part ...")
+        def getc(size, timeout=1):
+            return ser.read(size) or None
+
+        def putc(data, timeout=1):
+            bar.update(len(data));
+            bar.refresh()
+            return ser.write(data)  # note that this ignores the timeout
+
+        modem = XMODEM1k(getc, putc)
+
+        status = modem.send(f_stream, retry=100, quiet=True)
+
+        bar.close()
+    else:
+        def getc(size, timeout=1):
+            return ser.read(size) or None
+
+        def putc(data, timeout=1):
+            return ser.write(data)  # note that this ignores the timeout
+
+        modem = XMODEM1k(getc, putc)
+
+        status = modem.send(f_stream, retry=100, quiet=True)
+
+    close_serial_port()
+
+    return status
+
+#Receive the file from EVM
+def xmodem_recv_file(port, baudrate, timeout):
+
+    receiverespfilename = "receiverespfile.part"
+    recvrespfile = open(receiverespfilename, "wb")
+
+    open_serial_port();
+
+    def getc(size, timeout=1):
+        return ser.read(size) or None
+
+    def putc(data, timeout=1):
+        return ser.write(data)  # note that this ignores the timeout
+
+    modem = XMODEM1k(getc, putc)
+    status = modem.recv(recvrespfile, quiet=True, timeout=1000)
+
+    close_serial_port()
+
+    recvrespfile.close()
+
+    return receiverespfilename
+
+# Sends the mcelf image meta data, program segments as requested by EVM via xmodem
+# Receives request from EVM, parse the request frame and send the requested segment
+def xmodem_send_segments(filename, serialport, baudrate=115200, timeout=10):
+    status = False
+    timetaken = 0
+    show_progress=True
+
+    total_file_size = os.path.getsize(filename)
+    print(f"Total filesize = {total_file_size} bytes")
+
+    #Open the file
+    try:
+        stream = open(filename, 'rb')
+    except FileNotFoundError:
+        print('[ERROR] File [' + filename + '] not found !!!');
+        sys.exit();
+
+    #Read the file
+    file_bytes = stream.read()
+    stream.close()
+
+    stop_transaction = False
+
+    respfilename = "resp_cmd.dat"
+    resp_status = -1
+    offset = -1
+    req_length = -1
+
+    #Send the initial metadata of file of around 8kB
+    tstart = time.time()
+
+    with open("tempfilepart.part", 'wb') as f_part:
+        f_part.write(file_bytes[0:8192])
+    part_stream = open("tempfilepart.part", 'rb')
+    xmodem_send_file(serialport, baudrate, timeout, part_stream, show_progress)
+    part_stream.close()
+    os.remove("tempfilepart.part")
+
+    #Receive the request frame from target containing the offset and length of required segment
+    while(not stop_transaction):
+        resp_file_name = xmodem_recv_file(serialport, baudrate, 100)
+        resp_fh = open(resp_file_name, "rb")
+        buf = resp_fh.read(32)
+        resp_fh.close()
+        os.remove(resp_file_name)
+        resp_status, cmd, virt_mem_offset, req_length = parse_msg_response(buf)
+
+        if(resp_status==0):
+            if(cmd == BOOTLOADER_UART_SEND_FILE):
+                if(virt_mem_offset+req_length <= total_file_size):
+                    # Create a tempfile with the size requested
+                    with open("tempfilepart.part", 'wb') as f_part:
+                        f_part.write(file_bytes[virt_mem_offset:virt_mem_offset+req_length])
+                else:
+                # Error
+                   print(f"Request exceeding file size !! Sending remaining bytes ...")
+
+                part_stream = open("tempfilepart.part", 'rb')
+                xmodem_send_file(serialport, baudrate, timeout, part_stream, show_progress)
+                part_stream.close()
+                os.remove("tempfilepart.part")
+
+            elif(cmd == BOOTLOADER_UART_FILE_RECEIVE_COMPLETE):
+                resp_status = "[STATUS] Application load SUCCESS !!!"
+                stop_transaction = True
+            else:
+                resp_status = "[STATUS] ERROR: Application load FAILED !!!"
+        else:
+            pass
+
+    tstop = time.time()
+    timetaken = (tstop - tstart)
+
+    return resp_status, timetaken
 
 # Sends the file to EVM via xmodem, receives response from EVM and returns the response status
 def xmodem_send_receive_file(filename, serialport, baudrate=115200, get_response=True):
@@ -183,6 +358,11 @@ def main(argv):
         filecfg = FileCfg(config_file)
         parse_status = filecfg.parse()
 
+        if(filecfg.soc_index != None):
+            device = filecfg.cfgs[filecfg.soc_index].soc
+        else:
+            device = None
+
         if(parse_status != 0):
             print(parse_status)
             sys.exit()
@@ -192,9 +372,14 @@ def main(argv):
             print("Parsing config file ... SUCCESS. Found {} command(s) !!!".format(len(filecfg.cfgs)))
             print("")
 
+            if(filecfg.soc_index != None):
+                length = len(filecfg.cfgs) - 1
+            else:
+                length = len(filecfg.cfgs)
+
             if(filecfg.bootloader_index != None):
                 cfg_bootloader_file = filecfg.cfgs[filecfg.bootloader_index].bootloader
-                print("Executing command {} of {} ...".format(1, len(filecfg.cfgs)))
+                print("Executing command {} of {} ...".format(1, length))
                 print("Found the UART bootloader ... sending {}".format(cfg_bootloader_file))
                 send_status, timetaken = xmodem_send_receive_file(cfg_bootloader_file, serialport, get_response=False)
                 print("Sent bootloader {} of size {} bytes in {}s.".format(cfg_bootloader_file, os.path.getsize(cfg_bootloader_file), timetaken))
@@ -207,31 +392,42 @@ def main(argv):
 
             # loop through cfgs and lines, skip the process for bootloader
             for i in range(0, len(filecfg.cfgs)):
-                if(i != filecfg.bootloader_index):
+                if((i != filecfg.bootloader_index) and (i != filecfg.soc_index)):
                     line = filecfg.lines[i]
                     linecfg = filecfg.cfgs[i]
                     orig_filename = linecfg.filename
-                    print("Executing command {} of {} ...".format(i+1, len(filecfg.cfgs)))
+
+                    if(filecfg.soc_index != None):
+                        curr = i
+                    else:
+                        curr = i+1
+
+                    print("Executing command {} of {} ...".format(curr, length))
                     print("Command arguments : {}".format(line.rstrip('\n')))
-                    send_status, timetaken = xmodem_send_receive_file(linecfg.filename, serialport, get_response=True)
+
+                    if(device != "am275x"):
+                       send_status, timetaken = xmodem_send_receive_file(linecfg.filename, serialport, get_response=True)
+                    elif(device == "am275x"):
+                        send_status, timetaken = xmodem_send_segments(linecfg.filename, serialport)
+
                     print("Sent {} of size {} bytes in {}s.".format(orig_filename, os.path.getsize(orig_filename), timetaken))
                     print(send_status)
                     print("")
 
-
-            magic_word_filename = "magic_word_file.dat"
-            magic_word_file     = open(magic_word_filename,"wb")
-            magic_word_bytes    = BOOTLOADER_END_OF_FILES_TRANSFER.to_bytes(BOOTLOADER_END_OF_FILES_TRANSFER_WORD_LENGTH,"big")
-            magic_word_file.write(magic_word_bytes)
-            magic_word_file.close()
-            sendd_status, timetaken = xmodem_send_receive_file(magic_word_filename, serialport, get_response=False)
-            print("")
-            print(" Sent End Of File Transfer message of size {} bytes in {}s.".format( os.path.getsize(magic_word_filename), timetaken))
-            print("")
-            os.remove(magic_word_filename)
-            print("All commands from config file are executed !!!")
-            if("SUCCESS" in send_status):
-                print("Connect to UART in 5 seconds to see logs from UART !!!")
+            if(length != 1):
+                magic_word_filename = "magic_word_file.dat"
+                magic_word_file     = open(magic_word_filename,"wb")
+                magic_word_bytes    = BOOTLOADER_END_OF_FILES_TRANSFER.to_bytes(BOOTLOADER_END_OF_FILES_TRANSFER_WORD_LENGTH,"big")
+                magic_word_file.write(magic_word_bytes)
+                magic_word_file.close()
+                sendd_status, timetaken = xmodem_send_receive_file(magic_word_filename, serialport, get_response=False)
+                print("")
+                print(" Sent End Of File Transfer message of size {} bytes in {}s.".format( os.path.getsize(magic_word_filename), timetaken))
+                print("")
+                os.remove(magic_word_filename)
+                print("All commands from config file are executed !!!")
+                if("SUCCESS" in send_status):
+                    print("Connect to UART in 5 seconds to see logs from UART !!!")
 
     else:
         # Validate the cmdline config
@@ -276,11 +472,12 @@ def main(argv):
 
 # Class definitions used
 class LineCfg():
-    def __init__(self, line=None, filename=None, bootloader=None):
+    def __init__(self, line=None, filename=None, bootloader=None, soc=None):
         self.line = line
         self.filename = filename
         self.bootloader = bootloader
         self.exit_now = False
+        self.soc = soc
 
     # Takes a string of comma separated key=value pairs and parses into a dictionary
     def parse_to_dict(self, config_string):
@@ -313,7 +510,7 @@ class LineCfg():
                 status = "invalid_line"
             else:
                 # check for bootloader
-                if "--bootloader" not in config_dict.keys():
+                if ("--bootloader" not in config_dict.keys() and "--soc" not in config_dict.keys()):
                     if "--file" not in config_dict.keys():
                         status = "[ERROR] No filename provided !!!"
                         return status
@@ -326,7 +523,7 @@ class LineCfg():
                         status = "[ERROR] File not found !!!"
                         return status
 
-                else:
+                elif("--bootloader" in config_dict.keys()):
                     self.bootloader = config_dict["--bootloader"]
 
                     try:
@@ -334,6 +531,10 @@ class LineCfg():
                     except FileNotFoundError:
                         status = "[ERROR] Bootloader file not found !!!"
                         return status
+
+                elif("--soc" in config_dict.keys()):
+                    self.soc = config_dict["--soc"]
+                    return status
         else:
             # Should not hit this
             status = -1
@@ -345,6 +546,7 @@ class FileCfg():
         self.lines = list()
         self.cfgs = list()
         self.errors = list()
+        self.soc_index = None
         self.bootloader_index = None
 
     def parse(self):
@@ -375,6 +577,11 @@ class FileCfg():
                             self.lines.append(line.rstrip('\n'))
                             self.bootloader_index = valid_cfg_count
                             valid_cfg_count += 1
+                    elif(linecfg.soc != None):
+                        self.cfgs.append(linecfg)
+                        self.lines.append(line.rstrip('\n'))
+                        self.soc_index = valid_cfg_count
+                        valid_cfg_count += 1
                     else:
                         self.cfgs.append(linecfg)
                         self.lines.append(line.rstrip('\n'))
